+++ lib/modules/settings/modules/configuration_nzbget_headers/widgets.dart
@@ -1,2 +1,0 @@
-export 'widgets/add_header_tile.dart';
-export 'widgets/header_tile.dart';